Update google ads client to use campaign ids as-is / update tests
NOTES
=====
You are right that since the env var will be a comma separated string anyway,
we can just take in that. I updated main.py also to remove the split(,)
call, which should take care of the function receiving a list like
['']. We might want to do some splitting and trimming for validation sake
though.

Tested in GCP and it worked.

Change-Id: Ie33fa033c1ff2692911187268ed506fae5d632a1